made some progress
