+++ onboard/src/task_planning/task_planning/utils/other_utils.py
@@ -2,7 +2,8 @@ import os
 from collections.abc import Callable
 from enum import Enum
 from typing import Any
-
+from datetime import datetime
+import pytz
 
 class RobotName(Enum):
     """Enum for valid robot names."""
@@ -36,3 +37,24 @@ def singleton(cls: type) -> Callable[[Any, Any], type]:
             instances[cls] = cls(*args, **kwargs)
         return instances[cls]
     return getinstance
+
+def ros_timestamp_to_pacific_time(sec: int, nanosec: int) -> str:
+    """
+    Convert ROS timestamp (seconds and nanoseconds) to human-readable Pacific time.
+
+    Args:
+        sec (int): Seconds since epoch
+        nanosec (int): Nanoseconds
+
+    Returns:
+        str: Human-readable timestamp in Pacific timezone
+    """
+    # Convert to datetime object
+    pacific_tz = pytz.timezone('US/Pacific')
+    timestamp = datetime.fromtimestamp(sec + nanosec / 1e9, tz=pacific_tz)
+
+    # Convert to Pacific timezone
+    pacific_time = timestamp.astimezone(pacific_tz)
+
+    # Format as human-readable string
+    return pacific_time.strftime('%Y-%m-%d %H:%M:%S %Z')
